refactor: Cambio del tipo devuelto de int a long.
Se ha creado la necesidad de cambiar el tipo de dato devuelto de int a long, debido a que se ha cambiado a una nueva forma de calcular fibonacci.

+++ src/FibonacciWebApi/Controllers/FibonacciController.cs
@@ -13,10 +13,10 @@ namespace FibonacciWebApi.Controllers
     {
         // GET api/Fibonacci/5
         [HttpGet("{número}")]
-        public ActionResult<int> Get(int número)
+        public ActionResult<long> Get(int número)
         {
             Servicios ser = new Fibonacci();
-            int res = ser.CalcularFibonacci(número);
+            long res = ser.CalcularFibonacci(número);
             return res;
         }
     }
